# Refactor and optimize preferred (anti) pod affinity 